Hyperparameter tuning: Push to Git

+++ example_configs/speech2text/w2l_large_8gpus.py
@@ -43,7 +43,7 @@ base_params = {
   },
 
   #"max_grad_norm": 15.0,
-  "dtype": tf.float32,
+  "dtype": tf.float16,
 
   "summaries": ['learning_rate', 'variables', 'gradients', 'larc_summaries',
                 'variable_norm', 'gradient_norm', 'global_gradient_norm'],
